+++ resources/views/intro.blade.php
@@ -62,20 +62,9 @@
             <div class="row">
             	@foreach ($articles as $i=>$article)
 	                <div class="col-sm-6 col-xs-12 bottom-space">
-	                	<a href="{!! $article->getUrl() !!}" class="featured-article">
-	                		<img src="{!! $article->getThumbnailImage() !!}" class="img-responsive" alt="{!! $article->title !!}">
-	                	</a>
-                        <a href="{!! $article->getUrl() !!}"><h4 class="title">
-                            @if ($article->category)
-                                {!! $article->category->name !!}:
-                            @endif
-                            {!! $article->title !!}
-                        </h4></a>
-                        <p class="attributes">{!! $article->getShortTextAttribute($article->summary, 250) !!}
-                        (<a href="{!! $article->getUrl() !!}">{{ trans('general.more') }}</a>)
-                        </p>
-                        <p class="meta">{!!$article->published_date!!} / {!!$article->author!!}</p>
-	                    
+                        @include('components.article_thumbnail', [
+                            'article' => $article
+                        ])                    
 	                </div>
                     @if ($i%2 == 1)
                         <div class="clearfix"></div>