+++ resources/views/patternlib.blade.php
@@ -29,7 +29,6 @@ Pattern Library | @parent
             
             <section class="row">
                 <div class="col-xs-12">
-
                     <div class="panel panel-default">
                         <div class="panel-heading">
                             <h2 class="panel-title">{{$component['name']}}</h2>
@@ -54,6 +53,15 @@ Pattern Library | @parent
 
                             <h4>Usage notes</h4> 
                             <p>{{$component['usage_notes']}}</p>
+
+                            @if (isset($component['worked_on_this']))
+                                <h4>Worked on this</h4> 
+                                <p>
+                                    @foreach ($component['worked_on_this'] as $developer)
+                                        <span class="btn btn-default btn-xs btn-outline">{{ $developer }}</span>
+                                    @endforeach
+                                </p>
+                            @endif
 
                             <a href="#" class="btn btn-default btn-outline sans" onclick="toggle_source(event);"><i class="fa fa-code"></i> Show source</a>
                             <pre class="js-source pre-scrollable hidden"><code class="html">{{$component['source_code']}}</code></pre>