+++ tests/TestCase.php
@@ -29,4 +29,26 @@ class TestCase extends \Illuminate\Foundation\Testing\TestCase
 
         return $app;
     }
+
+    /**
+     * @param string $type
+     * @param string|null $message
+     * @param callable $function
+     */
+    protected function assertException($type, $message, callable $function)
+    {
+        $exception = null;
+
+        try {
+            call_user_func($function);
+        } catch (\Exception $e) {
+            $exception = $e;
+        }
+
+        self::assertThat($exception, new \PHPUnit_Framework_Constraint_Exception($type));
+
+        if ($message !== null) {
+            self::assertThat($exception, new \PHPUnit_Framework_Constraint_ExceptionMessage($message));
+        }
+    }
 }