+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+
+### Added
+- color-list component on artwork detail
+- Color description
+
 ### Changed
 - Favicons & Open Graph tags include components
 - positioned image reference strip on right side of zoom-viewer
@@ -41,6 +46,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [1.2.9] - 2017-12-05
 ### Changed
 - Enabled reproduction orders with mounting / framing
+- Componentized Article thumbnail
 
 ### Fixed
 - Database seeders